guardando datos v1

+++ TMEPortal/TMEPortal/DataContext/Payouts.Context.cs
@@ -35,7 +35,7 @@ namespace TMEPortal.DataContext
         public virtual DbSet<respuestaPagoMSI> respuestaPagoMSI { get; set; }
         public virtual DbSet<CFDVentaV33> CFDVentaV33 { get; set; }
     
-        public virtual int spMSIRespuestaPago(Nullable<int> moduloID, string mov, string movid, Nullable<int> sucursal, string cliente, string nombreCliente, string cp, string referencia, Nullable<System.DateTime> fechaRegistro, Nullable<decimal> importeTotal, Nullable<int> msi, Nullable<int> last4, Nullable<int> mesExp, Nullable<int> anioExp, string tipo)
+        public virtual int spMSIRespuestaPago(Nullable<int> moduloID, string mov, string movid, Nullable<int> sucursal, string cliente, string nombreCliente, string cp, string referencia, Nullable<System.DateTime> fechaRegistro, Nullable<decimal> importeTotal, Nullable<int> msi, Nullable<int> last4, Nullable<long> mesExp, Nullable<long> anioExp, string tipo)
         {
             var moduloIDParameter = moduloID.HasValue ?
                 new ObjectParameter("ModuloID", moduloID) :
